refactor: PG 추상화 누수 제거 — 결제 승인 요청 구성 책임을 각 PG 구현체로 이동
