+++ src/main/java/com/sesac/carematching/transaction/TransactionService.java
@@ -136,40 +136,9 @@ public class TransactionService {
             throw new IllegalStateException("결제 수단(PG사)이 선택되지 않았습니다. selectPg API를 먼저 호출해주세요.");
         }
 
-        // 현재 결제의 PG사에 알맞는 confirmRequestDTO 생성
-        PaymentConfirmRequestDTO request;
-        TransactionDetailDTO transactionDetailDTO;
         PaymentService paymentService = paymentServiceFactory.getService(pg);
-
-        if (pg == PaymentProvider.TOSS) {
-            // 토스페이먼츠는 자체적으로 결제 가격 확인 과정 추가
-            if (!transactionConfirmDTO.getPrice().equals(transaction.getPrice())) {
-                transaction.changeTransactionStatus(TransactionStatus.FAILED);
-                throw new IllegalStateException("잘못된 금액이 결제되었습니다. 다시 주문 해주세요.");
-            }
-            request = PaymentConfirmRequestDTO.builder()
-                .orderId(orderId)
-                .amount(transaction.getPrice())
-                .paymentKey(paymentKey)
-                .build();
-        } else if (pg == PaymentProvider.KAKAO) {
-            // 카카오는 추가 정보 필요
-            String pgToken = transactionConfirmDTO.getPgToken();
-            if (pgToken == null) {
-                throw new IllegalArgumentException("KakaoPay 결제시 pgToken은 필수값입니다.");
-            }
-            request = PaymentConfirmRequestDTO.builder()
-                .orderId(orderId)
-                .amount(transaction.getPrice())
-                .paymentKey(paymentKey)
-                .partnerUserId(userId.toString()) // 카카오: userId 추가 필요
-                .pgToken(pgToken) // 카카오: pgToken 추가 필요
-                .build();
-        } else {
-            throw new UnsupportedOperationException("confirm이 지원되지 않는 PG사: " + pg);
-        }
-
-        transactionDetailDTO = paymentService.confirmPayment(request);
+        PaymentConfirmRequestDTO request = paymentService.buildConfirmRequest(transaction, transactionConfirmDTO, paymentKey);
+        TransactionDetailDTO transactionDetailDTO = paymentService.confirmPayment(request);
         // DONE: 인증된 결제수단으로 요청한 결제가 승인된 상태입니다. (https://docs.tosspayments.com/reference#payment-%EA%B0%9D%EC%B2%B4)
         // KakaoPay 응답도 승인 성공시 자체적으로 Status를 DONE으로 설정하였음
         if (transactionDetailDTO.getPgStatus() != PgStatus.DONE)
