+++ src/main/java/com/sesac/carematching/transaction/payment/PaymentService.java
@@ -1,13 +1,14 @@
 package com.sesac.carematching.transaction.payment;
 
+import com.sesac.carematching.transaction.Transaction;
 import com.sesac.carematching.transaction.dto.PaymentReadyRequestDTO;
 import com.sesac.carematching.transaction.dto.PaymentConfirmRequestDTO;
 import com.sesac.carematching.transaction.dto.PaymentReadyResponseDTO;
+import com.sesac.carematching.transaction.dto.TransactionConfirmDTO;
 import com.sesac.carematching.transaction.dto.TransactionDetailDTO;
 
 /**
  * 결제 제공자(예: TossPayments, 다른 PG)들을 추상화한 서비스 인터페이스.
- * 각 PG사에서 지원하지 않는 메서드가 존재할 경우, 구체 클래스에서 UOE를 던져야 함.
  */
 public interface PaymentService {
 
@@ -47,5 +48,24 @@ public interface PaymentService {
      * @param request 주문 승인 정보
      */
     void healthCheckConfirm(PaymentConfirmRequestDTO request);
+
+    /**
+     * PG사에 맞는 결제 승인 요청 DTO를 구성합니다.
+     * 각 PG사마다 필요한 필드가 다르므로, 구현체가 직접 요청을 구성합니다.
+     *
+     * @param transaction  결제 대상 트랜잭션 엔티티
+     * @param clientInput  클라이언트(프론트엔드)에서 전달된 요청 데이터
+     * @param paymentKey   PG사에서 발급한 결제 키
+     * @return 각 PG사에 맞게 구성된 PaymentConfirmRequestDTO
+     */
+    PaymentConfirmRequestDTO buildConfirmRequest(Transaction transaction, TransactionConfirmDTO clientInput, String paymentKey);
+
+    /**
+     * PendingPayment 재시도 시 사용할 결제 승인 요청 DTO를 구성합니다.
+     *
+     * @param transaction  재시도 대상 트랜잭션 엔티티 (PendingPayment 포함)
+     * @return 재시도용 PaymentConfirmRequestDTO
+     */
+    PaymentConfirmRequestDTO buildRetryConfirmRequest(Transaction transaction);
 
 }
